+++ crate_anon/crateweb/research/views.py
@@ -999,8 +999,7 @@ def edit_display(request: HttpRequest, query_id: str) -> HttpResponse:
     query = get_object_or_404(Query, user=request.user, id=query_id)
     display_fields = query.get_display_list()
     try:
-        with query.get_executed_cursor() as cursor:
-            fieldnames = get_fieldnames_from_cursor(cursor)
+        fieldnames = query.get_column_names()
     except DatabaseError as exception:
         query.audit(failed=True, fail_msg=str(exception))
         return render_bad_query(request, query, exception)
@@ -1016,8 +1015,7 @@ def save_display(request: HttpRequest, query_id: str) -> HttpResponse:
     query = get_object_or_404(Query, user=request.user, id=query_id)
     if request.method == 'POST':
         try:
-            with query.get_executed_cursor() as cursor:
-                fieldnames = get_fieldnames_from_cursor(cursor)
+            fieldnames = query.get_column_names()
         except DatabaseError as exception:
             query.audit(failed=True, fail_msg=str(exception))
             return render_bad_query(request, query, exception)
@@ -1087,8 +1085,7 @@ def render_resultcount(request: HttpRequest, query: Query) -> HttpResponse:
     if query is None:
         return render_missing_query(request)
     try:
-        with query.get_executed_cursor() as cursor:
-            rowcount = cursor.rowcount
+        rowcount = query.get_rowcount()
         query.audit(count_only=True, n_records=rowcount)
         context = {
             'rowcount': rowcount,
@@ -1311,18 +1308,13 @@ def render_resultset(request: HttpRequest,
     if query is None:
         return render_missing_query(request)
     try:
-        with query.get_executed_cursor() as cursor:
-            all_fieldnames = get_fieldnames_from_cursor(cursor)
-            rows_all_cols = cursor.fetchall()
-            rowcount = cursor.rowcount
-            query.audit(n_records=rowcount)
+        rows = query.get_display_rows()
+        fieldnames = query.get_display_column_names()
+        rowcount = query.get_rowcount()
+        query.audit(n_records=rowcount)
     except DatabaseError as exception:
         query.audit(failed=True, fail_msg=str(exception))
         return render_bad_query(request, query, exception)
-    # Get subset of columns that the user wants to display
-    rows = query.get_display_columns_rows(rows_all_cols)
-    field_indexes = query.get_display_indexes()
-    fieldnames = [all_fieldnames[i] for i in field_indexes]
     row_indexes = list(range(len(rows)))
     # We don't need to process all rows before we paginate.
     page = paginate(request, row_indexes)
@@ -1393,18 +1385,13 @@ def render_resultset_recordwise(request: HttpRequest,
     if query is None:
         return render_missing_query(request)
     try:
-        with query.get_executed_cursor() as cursor:
-            all_fieldnames = get_fieldnames_from_cursor(cursor)
-            rows_all_cols = cursor.fetchall()
-            rowcount = cursor.rowcount
-            query.audit(n_records=rowcount)
+        rows = query.get_display_rows()
+        fieldnames = query.get_display_column_names()
+        rowcount = query.get_rowcount()
+        query.audit(n_records=rowcount)
     except DatabaseError as exception:
         query.audit(failed=True, fail_msg=str(exception))
         return render_bad_query(request, query, exception)
-    # Get subset of columns that the user wants to display
-    rows = query.get_display_columns_rows(rows_all_cols)
-    field_indexes = query.get_display_indexes()
-    fieldnames = [all_fieldnames[i] for i in field_indexes]
     row_indexes = list(range(len(rows)))
     # We don't need to process all rows before we paginate.
     page = paginate(request, row_indexes, per_page=1)